% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adtteSpec.R
\name{adtteSpecInput}
\alias{adtteSpecInput}
\title{Module Input for \code{ADTTE} Specification}
\usage{
adtteSpecInput(inputId, label_paramcd = "Select Endpoint")
}
\arguments{
\item{inputId}{(\code{string})\cr
the ID used to call the module input.}

\item{label_paramcd}{(\code{string})\cr label for the endpoint (\code{PARAMCD}) selection.}
}
\value{
The UI part.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This defines the input for the \code{ADTTE} specification.
}
\seealso{
\code{\link[=adtteSpecServer]{adtteSpecServer()}} for the module server and a complete example.
}
