% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assaySpec.R
\name{assaySpecInput}
\alias{assaySpecInput}
\title{Module Input for Assay Specification}
\usage{
assaySpecInput(inputId, label_assays = "Select Assay")
}
\arguments{
\item{inputId}{(\code{string})\cr
the ID used to call the module input.}

\item{label_assays}{(\code{string})\cr label for the assay selection.}
}
\value{
The UI part.
}
\description{
This defines the input for the assay specification.
}
\seealso{
\code{\link[=assaySpecServer]{assaySpecServer()}} for the module server and a complete example.
}
