% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assaySpec.R
\name{assaySpecServer}
\alias{assaySpecServer}
\title{Module Server for Assay Specification}
\usage{
assaySpecServer(inputId, assays, exclude_assays = character())
}
\arguments{
\item{inputId}{(\code{string})\cr
the ID used to call the module input.}

\item{assays}{(reactive \code{character})\cr available assays in the currently selected experiment.}

\item{exclude_assays}{(\code{character})\cr
names of the assays which should not be included in choices in the teal module.}
}
\value{
The chosen assay as a reactive string.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This defines the server part for the assay specification.
}
\seealso{
\code{\link[=assaySpecInput]{assaySpecInput()}} for the module UI.
}
