% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{assert_adtte_vars}
\alias{assert_adtte_vars}
\title{Check for \code{ADTTE} Variables}
\usage{
assert_adtte_vars(x)
}
\arguments{
\item{x}{an object to check.}
}
\description{
Check whether \code{x} is a list of \code{ADTTE} variables.
}
\examples{
assert_adtte_vars(list(aval = "AV", is_event = "EV", paramcd = "PC", usubjid = "ID", avalu = "u"))
}
\seealso{
\code{\link{assertions}} for more details.
}
