% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{assert_summary_funs}
\alias{assert_summary_funs}
\title{Check for List of Summary Functions}
\usage{
assert_summary_funs(x, null.ok = FALSE)
}
\arguments{
\item{x}{an object to check.}

\item{null.ok}{(\code{flag})\cr whether \code{x} may also contain \code{NULL}, meaning that
a user choice is possible where no summary function should be applied.}
}
\description{
Check whether \code{x} is a list of summary functions.
}
\examples{
assert_summary_funs(list(mean = colMeans, raw = NULL), null.ok = TRUE)
}
\seealso{
\code{\link{assertions}} for more details.
}
