% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_convention.R
\name{assertion_arguments}
\alias{assertion_arguments}
\title{Standard Assertion Arguments}
\arguments{
\item{x}{an object to check.}

\item{null.ok}{(\code{flag})\cr whether \code{x} may also be \code{NULL}.}

\item{.var.name}{(\code{string})\cr name of the checked object to print in
assertions; defaults to the heuristic implemented in \code{\link[checkmate:vname]{checkmate::vname()}}.}

\item{add}{(\code{AssertCollection} or \code{NULL})\cr collection to store
assertion messages, see \code{\link[checkmate:AssertCollection]{checkmate::AssertCollection}}.}
}
\description{
The documentation to this function lists all the conventional arguments in
additional \code{checkmate} assertions.
}
