% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{assertions}
\alias{assertions}
\title{Additional Assertions for \code{checkmate}}
\value{
Depending on the function prefix.
\itemize{
\item \code{assert_} functions return the object invisibly if successful, and otherwise
throw an error message.
\item \code{check_} functions return \code{TRUE} if successful, otherwise a string with the
error message.
\item \code{test_} functions just return \code{TRUE} or \code{FALSE}.
}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

We provide additional assertion functions which can be used together with
the \code{checkmate} functions. These are described in individual help pages
linked below.
}
\seealso{
\code{\link[=check_tag]{check_tag()}}
}
