% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_reactive}
\alias{check_reactive}
\alias{assert_reactive}
\alias{test_reactive}
\title{Check for Reactive Input}
\usage{
check_reactive(x)

assert_reactive(x, .var.name = checkmate::vname(x), add = NULL)

test_reactive(x)
}
\arguments{
\item{x}{an object to check.}

\item{.var.name}{(\code{string})\cr name of the checked object to print in
assertions; defaults to the heuristic implemented in \code{\link[checkmate:vname]{checkmate::vname()}}.}

\item{add}{(\code{AssertCollection} or \code{NULL})\cr collection to store
assertion messages, see \code{\link[checkmate:AssertCollection]{checkmate::AssertCollection}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Check whether \code{x} is a reactive input.
}
\examples{
check_reactive("bla")
check_reactive(reactive("bla"))
}
\seealso{
\code{\link{assertions}} for more details.
}
