% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_tag}
\alias{check_tag}
\alias{assert_tag}
\alias{test_tag}
\title{Check for Shiny Tag}
\usage{
check_tag(x, null.ok = FALSE)

assert_tag(x, null.ok = FALSE, .var.name = checkmate::vname(x), add = NULL)

test_tag(x, null.ok = FALSE)
}
\arguments{
\item{x}{an object to check.}

\item{null.ok}{(\code{flag})\cr whether \code{x} may also be \code{NULL}.}

\item{.var.name}{(\code{string})\cr name of the checked object to print in
assertions; defaults to the heuristic implemented in \code{\link[checkmate:vname]{checkmate::vname()}}.}

\item{add}{(\code{AssertCollection} or \code{NULL})\cr collection to store
assertion messages, see \code{\link[checkmate:AssertCollection]{checkmate::AssertCollection}}.}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Check whether \code{x} is a shiny tag.
}
\examples{
check_tag("bla")
check_tag(NULL, null.ok = TRUE)
}
\seealso{
\code{\link{assertions}} for more details.
}
