% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentSpec.R
\name{experimentSpecServer}
\alias{experimentSpecServer}
\title{Module Server for Experiment Specification}
\usage{
experimentSpecServer(
  inputId,
  datasets,
  mae_name,
  sample_vars_as_factors = TRUE,
  with_mae_col_data = TRUE
)
}
\arguments{
\item{inputId}{(\code{string})\cr
the ID used to call the module input.}

\item{datasets}{(\code{Datasets})\cr
teal specific argument which is automatically passed to the UI and server
functions, holding all the data sets provided in the app initialization.}

\item{mae_name}{(\code{string})\cr
name of the MAE data used in the teal module.}

\item{sample_vars_as_factors}{(\code{flag})\cr whether to convert the sample variables
(columns in \code{colData()} of the experiment) from character to factor variables.}

\item{with_mae_col_data}{(\code{flag})\cr whether to include the \code{colData()} of the
MAE into the experiment \code{colData()}.}
}
\value{
List with the following reactive objects:
\itemize{
\item \code{data}: the \code{\link[hermes:HermesData-class]{hermes::AnyHermesData}} experiment.
\item \code{name}: the name of the experiment as selected by the user.
\item \code{genes}: the names of the genes in \code{data}.
\item \code{assays}: the names of the assays in \code{data}.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This defines the server part for the experiment specification.
}
\examples{
todo
}
\seealso{
\code{\link[=experimentSpecInput]{experimentSpecInput()}} for the module UI.
}
