% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_convention.R
\name{function_arguments}
\alias{function_arguments}
\title{Standard Function Arguments}
\arguments{
\item{adtte}{(\verb{data frame})\cr an \code{adtte} dataset.}

\item{usubjid_var}{(\code{string})\cr variable name of the subject ID variable.}

\item{mae}{(\code{MultiAssayExperiment})\cr contains \code{AnyHermesData} objects.}

\item{object}{(\code{AnyHermesData})\cr contains RNA-seq values for one experiment.}

\item{genes}{(\code{GeneSpec})\cr specification for gene(s) (signature), e.g.
using \code{\link[hermes:gene_spec]{hermes::gene_spec()}}.}

\item{experiment_name}{(\code{string})\cr the desired \code{HermesData} to use.}

\item{assay_name}{(\code{string})\cr the assay to define the groups.}
}
\description{
The documentation to this function lists all the conventional arguments in
functions.
}
\keyword{internal}
