% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSpec.R
\name{geneSpecInput}
\alias{geneSpecInput}
\title{Module Input for Gene Signature Specification}
\usage{
geneSpecInput(
  inputId,
  funs,
  label_genes = "Select gene(s)",
  label_funs = "Select gene summary",
  label_text_button = "Enter list of genes",
  label_lock_button = "Lock gene selection"
)
}
\arguments{
\item{inputId}{(\code{string})\cr
the ID used to call the module input.}

\item{funs}{(named \code{list})\cr names of this list will be used for the function
selection drop down menu.}

\item{label_genes}{(\code{string})\cr label for the gene selection.}

\item{label_funs}{(\code{string})\cr label for the function selection.}

\item{label_text_button}{(\code{string})\cr label for the text input button.}

\item{label_lock_button}{(\code{string})\cr label for the lock button.}
}
\value{
The UI part.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This defines the input for the gene signature specification.
}
\examples{
geneSpecInput("my_genes", list(mean = colMeans), label_funs = "Please select function")
}
\seealso{
\code{\link[=geneSpecServer]{geneSpecServer()}} for the module server and a complete example.
}
