% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleVarSpec.R
\name{h_collapse_levels}
\alias{h_collapse_levels}
\title{Helper Function for Collapsing of Factor Levels}
\usage{
h_collapse_levels(x, group_list)
}
\arguments{
\item{x}{(\code{factor})\cr original factor.}

\item{group_list}{(named \code{list} of \code{character})\cr includes the collapsing
specification.}
}
\value{
The transformed factor \code{x} with new levels.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Given a group list and a factor, this helper function collapses the
levels in the factor accordingly and also ensures that the resulting
levels are in the order given in the group list.
}
\examples{
set.seed(123)
x <- factor(sample(
  c("ASIAN", "BLACK OR AFRICAN AMERICAN", "MULTIPLE", "UNKNOWN", "WHITE"),
  size = 30L,
  replace = TRUE
))
group_list <- list(
  "ASIAN/BLACK OR AFRICAN AMERICAN" = c("ASIAN", "BLACK OR AFRICAN AMERICAN"),
  "MULTIPLE/UNKNOWN" = c("MULTIPLE", "UNKNOWN"),
  "WHITE" = "WHITE"
)
x_collapsed <- h_collapse_levels(x, group_list)
stopifnot(identical(levels(x_collapsed), names(group_list)))
}
