% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_extract_words}
\alias{h_extract_words}
\title{Helper Function to Extract Words}
\usage{
h_extract_words(x)
}
\arguments{
\item{x}{(\code{string})\cr input.}
}
\value{
Character vector with the extracted words.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This helper function extracts words from a string. Here words are defined
as containing lower or upper case letters, colons and dots. All other
characters are considered separators.
}
\examples{
h_extract_words("a, b, , c, 234; 34562 - GeneID:bla")
h_extract_words("GeneID:1820, sdf.393; 32596")
}
