% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentSpec.R
\name{h_gene_data}
\alias{h_gene_data}
\title{Helper Function to Format Gene Choices}
\usage{
h_gene_data(object, name_annotation)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr contains RNA-seq values for one experiment.}

\item{name_annotation}{(\code{string} or \code{NULL})\cr which annotation column to use as name
to return in the \code{genes} data. If \code{NULL}, then the \code{name} column will be set to empty
strings.}
}
\value{
A \code{data.frame} with \code{id} and \code{name} columns containing all genes from
\code{object}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Given a \code{\link[hermes:HermesData-class]{hermes::AnyHermesData}} data object, as well as the annotation
column name to use as gene name, this function formats the contained genes
as a \code{data.frame} ready for consumption in \code{\link[=h_order_genes]{h_order_genes()}} e.g.
}
\details{
Note that missing names or names that only contain whitespace
are replaced by empty strings for consistency and better labeling in the
UI downstream
}
\examples{
object <- hermes::hermes_data[1:10, ]
h_gene_data(object, "symbol")
}
