% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentSpec.R
\name{h_order_genes}
\alias{h_order_genes}
\title{Helper Function to Order Gene Choices}
\usage{
h_order_genes(genes)
}
\arguments{
\item{genes}{(\code{data.frame})\cr containing \code{id} and \code{name} columns of the
gene choices. Note that no missing values are allowed.}
}
\value{
The ordered \code{data.frame}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The possible gene choices are ordered as follows. First come all genes which
have a non-empty name, ordered by their name alphabetically. Last come
all genes with an empty name, ordered by their ID alphabetically.
}
\examples{
genes <- data.frame(
  id = c("7", "1", "2", "345346", "0"),
  name = c("e", "", "c", "", "a")
)
h_order_genes(genes)
}
