% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSpec.R
\name{h_parse_genes}
\alias{h_parse_genes}
\title{Helper Function to Parse Genes}
\usage{
h_parse_genes(words, choices)
}
\arguments{
\item{words}{(\code{character})\cr containing gene IDs or names.}

\item{choices}{(\code{data.frame})\cr containing \code{id} and \code{name} columns of the
new choices.}
}
\value{
The subset of \code{choices} which matches \code{words} in ID or name.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This helper function takes a vector of \code{words} and tries to match them
with the \code{id} and \code{name} columns of possible gene choices.
}
\examples{
h_parse_genes(
  c("a", "2535"),
  data.frame(id = as.character(2533:2537), name = letters[1:5])
)
}
