% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSpec.R
\name{h_update_gene_selection}
\alias{h_update_gene_selection}
\title{Helper Function to Update Gene Selection}
\usage{
h_update_gene_selection(session, inputId, selected, choices)
}
\arguments{
\item{session}{(\code{ShinySession})\cr the session object.}

\item{inputId}{(\code{string})\cr
the ID used to call the module input.}

\item{selected}{(\code{character})\cr currently selected gene IDs.}

\item{choices}{(\code{data.frame})\cr containing \code{id} and \code{name} columns of the
new choices.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This helper function takes the intersection of \code{selected} and
\code{choices} for genes and updates the \code{inputId} accordingly. It then
shows a notification if not all \code{selected} genes were available.
}
