% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_quality.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{Correlation Heatmap Plot}
\usage{
heatmap_plot(object, assay_name)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr contains RNA-seq values for one experiment.}

\item{assay_name}{(\code{string})\cr the assay to define the groups.}
}
\value{
Plot to be displayed in the teal app.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function plots the correlation heatmap.
}
\examples{
library(hermes)
object <- HermesData(summarized_experiment)
result <- heatmap_plot(object, assay_name = "counts")
}
