% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_convention.R
\name{module_arguments}
\alias{module_arguments}
\title{Standard Module Arguments}
\arguments{
\item{datasets}{(\code{Datasets})\cr
teal specific argument which is automatically passed to the UI and server
functions, holding all the data sets provided in the app initialization.}

\item{label}{(\code{string})\cr
menu item label of the module in the teal app.}

\item{id}{(\code{string})\cr
the ID used to call the teal module's UI function.}

\item{mae_name}{(\code{string})\cr
name of the MAE data used in the teal module.}

\item{exclude_assays}{(\code{character})\cr
names of the assays which should not be included in choices in the teal module.}

\item{pre_output}{(\code{shiny.tag} or \code{NULL})\cr
placed before the output to put the output into context (for example a title).}

\item{post_output}{(\code{shiny.tag} or \code{NULL})\cr
placed after the output to put the output into context (for example the \code{\link[shiny:helpText]{shiny::helpText()}}
elements can be useful).}

\item{input}{(\code{ReactiveValues})\cr the session's input object.}

\item{output}{(\code{shinyoutput})\cr the session's output object.}

\item{session}{(\code{ShinySession})\cr the session object.}
}
\description{
The documentation to this function lists all the conventional arguments in
\code{hermes} teal modules.
}
