% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleVarSpec.R
\name{sampleVarSpecInput}
\alias{sampleVarSpecInput}
\title{Module Input for Sample Variable Specification}
\usage{
sampleVarSpecInput(
  inputId,
  label_vars = "Select sample variable",
  label_levels_button = "Combine factor levels"
)
}
\arguments{
\item{inputId}{(\code{string})\cr
the ID used to call the module input.}

\item{label_vars}{(\code{string})\cr label for the sample variable selection.}

\item{label_levels_button}{(\code{string})\cr label for the levels combination button.}
}
\value{
The UI part.
}
\description{
This defines the input for the sample variable specification.
}
\examples{
sampleVarSpecInput("my_vars", label_vars = "Select faceting variable")
}
\seealso{
\code{\link[=sampleVarSpecServer]{sampleVarSpecServer()}} for the module server and a complete example.
}
