% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleVarSpec.R
\name{sampleVarSpecServer}
\alias{sampleVarSpecServer}
\title{Module Server for Sample Variable Specification}
\usage{
sampleVarSpecServer(
  id,
  experiment_name,
  original_data,
  transformed_data = original_data,
  assign_lists = reactiveValues(),
  num_levels = NULL,
  label_modal_title = "Please click to group the original factor levels"
)
}
\arguments{
\item{id}{(\code{string}) the shiny module id.}

\item{experiment_name}{(reactive \code{string})\cr name of the input experiment.}

\item{original_data}{(reactive \code{SummarizedExperiment})\cr input experiment where the
sample variables extracted via \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::colData()}} should be eligible for
selection.}

\item{transformed_data}{(reactive \code{SummarizedExperiment})\cr used when multiple sample
variables can be selected in the app. In that case, pass here the pre-transformed data.}

\item{assign_lists}{(\code{reactivevalues})\cr object to share factor level groupings across multiple
sample variables.}

\item{num_levels}{(\code{count} or \code{NULL})\cr required number of levels after combining original levels.
If \code{NULL} then all numbers of levels are allowed.}

\item{label_modal_title}{(\code{string})\cr title for the dialog that asks for the text input.}
}
\value{
Reactive \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}} which can be used as
input for the relevant \code{hermes} functions.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This defines the server part for the sample variable specification.
}
\note{
Only atomic columns (e.g. not \code{DataFrame} columns) of the \code{colData}
which are not completely missing (\code{NA}) will be shown for selection.
If \code{num_levels} is specified then only factor columns will be available.
}
\examples{
ui <- function(id,
               datasets) {
  ns <- NS(id)
  mae <- datasets$get_data("MAE", filtered = FALSE)
  experiment_name_choices <- names(mae)
  teal.devel::standard_layout(
    encoding = div(
      selectInput(ns("experiment_name"), "Select experiment", experiment_name_choices),
      sampleVarSpecInput(ns("facet_var"), "Select faceting variable")
    ),
    output = plotOutput(ns("plot"))
  )
}
server <- function(id,
                   datasets) {
  moduleServer(id, function(input, output, session) {
    experiment_data <- reactive({
      req(input$experiment_name)
      mae <- datasets$get_data("MAE", filtered = TRUE)
      object <- mae[[input$experiment_name]]
      SummarizedExperiment::colData(object) <- hermes::df_cols_to_factor(SummarizedExperiment::colData(object))
      object
    })
    facet_var_spec <- sampleVarSpecServer(
      "facet_var",
      experiment_name = reactive({
        input$experiment_name
      }),
      original_data = experiment_data
    )
    output$plot <- renderPlot({
      experiment_data_final <- facet_var_spec$experiment_data()
      facet_var <- facet_var_spec$sample_var()
      hermes::draw_boxplot(
        experiment_data_final,
        assay_name = "counts",
        genes = hermes::genes(experiment_data_final)[1],
        facet_var = facet_var
      )
    })
  })
}
my_app <- function() {
  mae <- hermes::multi_assay_experiment
  mae_data <- dataset("MAE", mae)
  data <- teal_data(mae_data)
  app <- init(
    data = data,
    modules = root_modules(
      module(
        label = "sampleVarSpec example",
        server = server,
        ui = ui,
        filters = "all"
      )
    )
  )
  shinyApp(app$ui, app$server)
}
if (interactive()) {
  my_app()
}
}
\seealso{
\code{\link[=sampleVarSpecInput]{sampleVarSpecInput()}} for the module UI.
}
