% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{teal.modules.hermes-package}
\alias{teal.modules.hermes}
\alias{teal.modules.hermes-package}
\title{\code{teal.modules.hermes} Package}
\description{
\code{teal.modules.hermes} provides RNA-seq analysis modules to add to a
teal application.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://insightsengineering.github.io/teal.modules.hermes/}
  \item \url{https://github.com/insightsengineering/teal.modules.hermes/}
  \item Report bugs at \url{https://github.com/insightsengineering/teal.modules.hermes/issues}
}

}
\author{
\strong{Maintainer}: Daniel Sabanés Bové \email{daniel.sabanes_bove@roche.com}

Authors:
\itemize{
  \item Nikolas Burkoff
  \item Dinakar Kulkarni
  \item Konrad Pagacz
  \item Namrata Bhatia
  \item Dawid Kaledkowski
  \item Pawel Rucki
  \item Jeff Luong
  \item Stefanie Bienert
  \item Haocheng Li
  \item Lyndsee Midori Zhang
}

Other contributors:
\itemize{
  \item Sorin Voicu [contributor]
  \item Benoit Falquet [contributor]
  \item Mahmoud Hallal [contributor]
  \item Tim Treis [contributor]
  \item Vedha Viyash [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
