% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{template_g_km}
\alias{template_g_km}
\title{Template: Kaplan-Meier}
\usage{
template_g_km(
  dataname = "ANL",
  arm_var = "GENE",
  quantiles = c(0.33, 0.66),
  ref_arm = NULL,
  comp_arm = NULL,
  compare_arm = FALSE,
  combine_comp_arms = FALSE,
  aval_var = "AVAL",
  cnsr_var = "CNSR",
  xticks = NULL,
  strata_var = NULL,
  time_points = NULL,
  facet_var = "SEX",
  font_size = 8,
  conf_level = 0.95,
  ties = "efron",
  xlab = "Survival time",
  time_unit_var = "AVALU",
  yval = "Survival",
  pval_method = "log-rank",
  annot_surv_med = TRUE,
  annot_coxph = TRUE,
  ci_ribbon = FALSE,
  title = "KM Plot"
)
}
\arguments{
\item{xticks}{(\code{numeric}, \code{number}, or \code{NULL})\cr
numeric vector of ticks or single number with spacing
between ticks on the x axis.
If \code{NULL} (default), \code{\link[labeling:extended]{labeling::extended()}} is used to determine
an optimal tick position on the x axis.}

\item{facet_var}{(\code{character})\cr
object with all available choices and preselected option
for variable names that can be used for facet plotting.}

\item{font_size}{(\code{number}) \cr font size to be used.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{ties}{(\code{string})\cr among \code{exact} (equivalent to \code{DISCRETE} in SAS),
\code{efron} and \code{breslow}, see \code{\link[survival:coxph]{survival::coxph()}}.
Note: there is no equivalent of SAS \code{EXACT} method in R.}

\item{xlab}{(\code{string})\cr label of x-axis.}

\item{yval}{(\code{string})\cr value of y-axis. Should be either
\code{Survival} (default) or \code{Failure} probability.}

\item{pval_method}{(\code{string})\cr the method used for estimation of p.values;
\code{wald} (default) or \code{likelihood}.}

\item{annot_surv_med}{(\code{flag})\cr compute and add the annotation table
on the Kaplan-Meier curve estimating the median survival time per group.}

\item{annot_coxph}{(\code{flag})\cr add the annotation table from a \code{\link[survival:coxph]{survival::coxph()}} model.}

\item{ci_ribbon}{(\code{flag})\cr draw the confidence interval around the Kaplan-Meier curve.}

\item{title}{(\code{string})\cr title for plot.}
}
\description{
Template: Kaplan-Meier
}
\seealso{
\code{\link[=tm_g_km]{tm_g_km()}}
}
