% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_barplot.R
\name{tm_g_barplot}
\alias{tm_g_barplot}
\alias{ui_g_barplot}
\alias{srv_g_barplot}
\alias{sample_tm_g_barplot}
\title{Teal Module for RNA-seq Barplot}
\usage{
tm_g_barplot(
  label,
  mae_name,
  exclude_assays = character(),
  summary_funs = list(Mean = colMeans, Median = matrixStats::colMedians, Max =
    matrixStats::colMaxs),
  pre_output = NULL,
  post_output = NULL,
  .test = FALSE,
  transformators = list()
)

ui_g_barplot(
  id,
  mae_name,
  summary_funs,
  pre_output,
  post_output,
  .test = FALSE
)

srv_g_barplot(
  id,
  data,
  filter_panel_api,
  reporter,
  mae_name,
  exclude_assays,
  summary_funs,
  .test = FALSE
)

sample_tm_g_barplot(.test = FALSE)
}
\arguments{
\item{label}{(\code{string})\cr
menu item label of the module in the teal app.}

\item{mae_name}{(\code{string})\cr
name of the MAE data used in the teal module.}

\item{exclude_assays}{(\code{character})\cr
names of the assays which should not be included in choices in the teal module.}

\item{summary_funs}{(named \code{list} of functions or \code{NULL})\cr functions which can be used
in the the gene signatures. For modules that support also multiple genes without
summary, \code{NULL} can be included to not summarize the genes but provide all of them.}

\item{pre_output}{(\code{shiny.tag} or \code{NULL})\cr
placed before the output to put the output into context (for example a title).}

\item{post_output}{(\code{shiny.tag} or \code{NULL})\cr
placed after the output to put the output into context (for example the \code{\link[shiny:helpText]{shiny::helpText()}}
elements can be useful).}

\item{.test}{(\code{flag})\cr whether to display the internal structure of the plot for testing purposes.}

\item{transformators}{(\code{list} of \code{teal_transform_module}) that will be applied to transform module's data input.
To learn more check \code{vignette("transform-input-data", package = "teal")}.}

\item{id}{(\code{string}) the shiny module id.}

\item{data}{(\code{reactive})\cr
\verb{reactive(<teal_data>)} holding all the data sets provided
during app initialization after going through the filters.}

\item{filter_panel_api}{(\code{FilterPanelAPI})\cr object describing the actual filter panel API.}

\item{reporter}{(\code{Reporter}) object}
}
\value{
Shiny module to be used in the teal app.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This module provides an interactive barplot for RNA-seq gene expression
analysis.
}
\section{Functions}{
\itemize{
\item \code{ui_g_barplot()}: sets up the user interface.

\item \code{srv_g_barplot()}: sets up the server with reactive graph.

\item \code{sample_tm_g_barplot()}: sample module function.

}}
\examples{
data <- teal_data(MAE = hermes::multi_assay_experiment)
app <- init(
  data = data,
  modules = modules(
    tm_g_barplot(
      label = "barplot",
      mae_name = "MAE"
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

# Alternatively you can run the sample module with this function call:
if (interactive()) {
  sample_tm_g_barplot()
}
}
