% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.R
\name{tm_g_forest_tte}
\alias{tm_g_forest_tte}
\alias{ui_g_forest_tte}
\alias{srv_g_forest_tte}
\alias{sample_tm_g_forest_tte}
\title{Teal Module for Survival Forest Plot}
\usage{
tm_g_forest_tte(
  label,
  mae_name,
  pre_output = NULL,
  post_output = NULL,
  plot_height = c(600L, 200L, 2000L),
  plot_width = c(1360L, 500L, 2000L)
)

ui_g_forest_tte(id, datasets, mae_name, pre_output, post_output)

srv_g_forest_tte(
  input,
  output,
  session,
  datasets,
  mae_name,
  plot_height,
  plot_width
)

sample_tm_g_forest_tte()
}
\arguments{
\item{label}{(\code{string})\cr
menu item label of the module in the teal app.}

\item{mae_name}{(\code{string})\cr
name of the MAE data used in the teal module.}

\item{pre_output}{(\code{shiny.tag} or \code{NULL})\cr
placed before the output to put the output into context (for example a title).}

\item{post_output}{(\code{shiny.tag} or \code{NULL})\cr
placed after the output to put the output into context (for example the \code{\link[shiny:helpText]{shiny::helpText()}}
elements can be useful).}

\item{plot_height}{(\code{list})\cr list of integers to set the default, minimum,
and maximum plot height.}

\item{plot_width}{(\code{list})\cr list of integers to set the default, minimum,
and maximum plot width.}

\item{id}{(\code{string})\cr
the ID used to call the teal module's UI function.}

\item{datasets}{(\code{Datasets})\cr
teal specific argument which is automatically passed to the UI and server
functions, holding all the data sets provided in the app initialization.}

\item{input}{(\code{ReactiveValues})\cr the session's input object.}

\item{output}{(\code{shinyoutput})\cr the session's output object.}

\item{session}{(\code{ShinySession})\cr the session object.}
}
\value{
Shiny module to be used in the teal app.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This module provides an interactive survival forest plot.
}
\section{Functions}{
\itemize{
\item \code{ui_g_forest_tte}: sets up the user interface.

\item \code{srv_g_forest_tte}: sets up the server with reactive graph.

\item \code{sample_tm_g_forest_tte}: sample module function.
}}

\examples{
mae <- hermes::multi_assay_experiment
mae_data <- dataset("MAE", mae)
data <- teal_data(mae_data)
app <- init(
  data = data,
  modules = root_modules(
    static = {
      tm_g_forest_tte(
        label = "forestplot",
        mae_name = "MAE"
      )
    }
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}
\dontrun{
# Alternatively you can run the sample module with this function call:
sample_tm_g_forest_tte()
}
}
