% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{tm_g_scatterplot}
\alias{tm_g_scatterplot}
\alias{ui_g_scatterplot}
\alias{srv_g_scatterplot}
\alias{sample_tm_g_scatterplot}
\title{Teal Module for RNA-seq Scatterplot}
\usage{
tm_g_scatterplot(label, mae_name, pre_output = NULL, post_output = NULL)

ui_g_scatterplot(id, datasets, mae_name, pre_output, post_output)

srv_g_scatterplot(input, output, session, datasets, mae_name)

sample_tm_g_scatterplot()
}
\arguments{
\item{label}{(\code{string})\cr
menu item label of the module in the teal app.}

\item{mae_name}{(\code{string})\cr
name of the MAE data used in the teal module.}

\item{pre_output}{(\code{shiny.tag} or \code{NULL})\cr
placed before the output to put the output into context (for example a title).}

\item{post_output}{(\code{shiny.tag} or \code{NULL})\cr
placed after the output to put the output into context (for example the \code{\link[shiny:helpText]{shiny::helpText()}}
elements can be useful).}

\item{id}{(\code{string})\cr
the ID used to call the teal module's UI function.}

\item{datasets}{(\code{Datasets})\cr
teal specific argument which is automatically passed to the UI and server
functions, holding all the data sets provided in the app initialization.}

\item{input}{(\code{ReactiveValues})\cr the session's input object.}

\item{output}{(\code{shinyoutput})\cr the session's output object.}

\item{session}{(\code{ShinySession})\cr the session object.}
}
\value{
Shiny module to be used in the teal app.
}
\description{
This module provides an interactive scatterplot for RNA-seq gene expression
analysis.
}
\section{Functions}{
\itemize{
\item \code{ui_g_scatterplot}: sets up the user interface.

\item \code{srv_g_scatterplot}: sets up the server with reactive graph.

\item \code{sample_tm_g_scatterplot}: sample module function.
}}

\examples{
library(hermes)
mae <- hermes::multi_assay_experiment
for (i in seq_along(mae)) {
  mae[[i]] <- hermes::HermesData(mae[[i]])
}
mae_data <- dataset("MAE", mae)
data <- teal_data(mae_data)
app <- init(
  data = data,
  modules = root_modules(
    static = {
      tm_g_scatterplot(
        label = "scatterplot",
        mae_name = "MAE"
      )
    }
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}
\dontrun{
# Alternatively you can run the sample module with this function call:
sample_tm_g_scatterplot()
}
}
