% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_volcanoplot.R
\name{tm_g_volcanoplot}
\alias{tm_g_volcanoplot}
\alias{ui_g_volcanoplot}
\alias{srv_g_volcanoplot}
\alias{sample_tm_g_volcanoplot}
\title{Teal Module for RNA-seq Volcano Plot}
\usage{
tm_g_volcanoplot(
  label,
  mae_name,
  exclude_assays = character(),
  pre_output = NULL,
  post_output = NULL,
  .test = FALSE,
  transformators = list()
)

ui_g_volcanoplot(id, mae_name, pre_output, post_output, .test = FALSE)

srv_g_volcanoplot(
  id,
  data,
  filter_panel_api,
  reporter,
  mae_name,
  exclude_assays,
  .test = FALSE
)

sample_tm_g_volcanoplot(.test = FALSE)
}
\arguments{
\item{label}{(\code{string})\cr
menu item label of the module in the teal app.}

\item{mae_name}{(\code{string})\cr
name of the MAE data used in the teal module.}

\item{exclude_assays}{(\code{character})\cr
names of the assays which should not be included in choices in the teal module.}

\item{pre_output}{(\code{shiny.tag} or \code{NULL})\cr
placed before the output to put the output into context (for example a title).}

\item{post_output}{(\code{shiny.tag} or \code{NULL})\cr
placed after the output to put the output into context (for example the \code{\link[shiny:helpText]{shiny::helpText()}}
elements can be useful).}

\item{.test}{(\code{flag})\cr whether to display the internal structure of the plot for testing purposes.}

\item{transformators}{(\code{list} of \code{teal_transform_module}) that will be applied to transform module's data input.
To learn more check \code{vignette("transform-input-data", package = "teal")}.}

\item{id}{(\code{string}) the shiny module id.}

\item{data}{(\code{reactive})\cr
\verb{reactive(<teal_data>)} holding all the data sets provided
during app initialization after going through the filters.}

\item{filter_panel_api}{(\code{FilterPanelAPI})\cr object describing the actual filter panel API.}

\item{reporter}{(\code{Reporter}) object}
}
\value{
Shiny module to be used in the teal app.
}
\description{
This module provides an interactive volcano plot for RNA-seq gene expression
analysis.
}
\section{Functions}{
\itemize{
\item \code{ui_g_volcanoplot()}: sets up the user interface.

\item \code{srv_g_volcanoplot()}: sets up the server with reactive graph.

\item \code{sample_tm_g_volcanoplot()}: sample module function.

}}
\examples{
data <- teal_data(MAE = hermes::multi_assay_experiment)
app <- init(
  data = data,
  modules = modules(
    tm_g_volcanoplot(
      label = "volcanoplot",
      mae_name = "MAE"
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

# Alternatively you can run the sample module with this function call:
if (interactive()) {
  sample_tm_g_volcanoplot()
}
}
