% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_g_quality.R
\name{top_gene_plot}
\alias{top_gene_plot}
\title{Most Expressed Genes Plot}
\usage{
top_gene_plot(object, assay_name)
}
\arguments{
\item{object}{(\code{AnyHermesData})\cr contains RNA-seq values for one experiment.}

\item{assay_name}{(\code{string})\cr the assay to define the groups.}
}
\value{
Plot to be displayed in the teal app.
}
\description{
This function plots the most expressed genes.
}
\examples{
library(hermes)
object <- HermesData(summarized_experiment)
result <- top_gene_plot(object, assay_name = "counts")
}
