% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSpec.R
\name{validate_gene_spec}
\alias{validate_gene_spec}
\title{Validation of Gene Specification}
\usage{
validate_gene_spec(gene_spec, gene_choices)
}
\arguments{
\item{gene_spec}{(\code{GeneSpec})\cr gene specification.}

\item{gene_choices}{(\code{data.frame})\cr all possible gene choices, with \code{id}
and \code{name} columns.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This validation function checks that a given \code{\link[hermes:GeneSpec]{hermes::GeneSpec}} has at least
one gene selected and that all genes are included in possible choices.
}
