% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleVarSpec.R
\name{validate_n_levels}
\alias{validate_n_levels}
\title{Validation of Number of Levels}
\usage{
validate_n_levels(x, name, n_levels)
}
\arguments{
\item{x}{(\code{factor})\cr factor to validate.}

\item{name}{(\code{string})\cr name of \code{x} in the app.}

\item{n_levels}{(\code{count})\cr required number of factor levels in \code{x}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This validation function checks that a given vector \code{x} is a factor with
the specified number of levels.
}
