% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{assertions}
\alias{assertions}
\title{Additional Assertions for \code{checkmate}}
\value{
Depending on the function prefix.
\itemize{
\item \code{assert_} functions return the object invisibly if successful, and otherwise
throw an error message.
\item \code{check_} functions return \code{TRUE} if successful, otherwise a string with the
error message.
\item \code{test_} functions just return \code{TRUE} or \code{FALSE}.
}
}
\description{
We provide additional assertion functions which can be used together with
the \code{checkmate} functions. These are described in individual help pages
linked below.
}
\seealso{
\code{\link[=assert_tag]{assert_tag()}}, \code{\link[=assert_reactive]{assert_reactive()}}, \code{\link[=assert_summary_funs]{assert_summary_funs()}}, \code{\link[=assert_adtte_vars]{assert_adtte_vars()}}
}
\keyword{internal}
