% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentSpec.R
\name{experimentSpecInput}
\alias{experimentSpecInput}
\title{Module Input for Experiment Specification}
\usage{
experimentSpecInput(
  inputId,
  datasets,
  mae_name,
  label_experiments = "Select Experiment"
)
}
\arguments{
\item{inputId}{(\code{string})\cr
the ID used to call the module input.}

\item{datasets}{(\code{Datasets})\cr
teal specific argument which is automatically passed to the UI and server
functions, holding all the data sets provided in the app initialization.}

\item{mae_name}{(\code{string})\cr
name of the MAE data used in the teal module.}

\item{label_experiments}{(\code{string})\cr label for the experiment selection.}
}
\value{
The UI part.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This defines the input for the experiment specification.
}
\seealso{
\code{\link[=experimentSpecServer]{experimentSpecServer()}} for the module server and a complete example.
}
