% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_convention.R
\name{function_arguments}
\alias{function_arguments}
\title{Standard Function Arguments}
\arguments{
\item{adtte}{(\verb{data frame})\cr an \code{adtte} dataset.}

\item{mae}{(\code{MultiAssayExperiment})\cr contains \code{AnyHermesData} objects.}

\item{object}{(\code{AnyHermesData})\cr contains RNA-seq values for one experiment.}

\item{gene_var}{(\code{character})\cr a gene(s) of interest.}

\item{experiment_name}{(\code{string})\cr the desired experiment to use.}

\item{assay_name}{(\code{string})\cr the assay to define the groups.}
}
\description{
The documentation to this function lists all the conventional arguments in
functions.
}
