% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSpec.R
\name{geneSpecInput}
\alias{geneSpecInput}
\title{Module Input for Gene Signature Specification}
\usage{
geneSpecInput(
  inputId,
  funs,
  label_genes = "Select Gene(s)",
  label_funs = "Select Gene Summary",
  label_text_button = "Enter list of genes",
 
    label_lock_button = "Lock gene selection (so that it does not get updated when filtering)",
  label_select_all_button = paste0("Select All Genes (first ", max_options, ")"),
  label_select_none_button = "Select None",
  max_options = 200L,
  max_selected = max_options
)
}
\arguments{
\item{inputId}{(\code{string})\cr
the ID used to call the module input.}

\item{funs}{(named \code{list})\cr names of this list will be used for the function
selection drop down menu.}

\item{label_genes}{(\code{string})\cr label for the gene selection.}

\item{label_funs}{(\code{string})\cr label for the function selection.}

\item{label_text_button}{(\code{string})\cr label for the text input button.}

\item{label_lock_button}{(\code{string})\cr label for the lock button.}

\item{label_select_all_button}{(\code{string})\cr label for the selecting all genes button.}

\item{label_select_none_button}{(\code{string})\cr label for the selecting no genes button.}

\item{max_options}{(\code{count})\cr maximum number of gene options rendering and selected via
"Select All".}

\item{max_selected}{(\code{count})\cr maximum number of genes which can be selected.}
}
\value{
The UI part.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This defines the input for the gene signature specification.
}
\examples{
geneSpecInput("my_genes", list(mean = colMeans), label_funs = "Please select function")
}
\seealso{
\code{\link[=geneSpecServer]{geneSpecServer()}} for the module server and a complete example.
}
