% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{h_km_mae_to_adtte}
\alias{h_km_mae_to_adtte}
\title{Data Preprocessing for KM Module}
\usage{
h_km_mae_to_adtte(
  adtte,
  mae,
  gene_var,
  experiment_name = "se1",
  assay_name = "counts"
)
}
\arguments{
\item{adtte}{(\verb{data frame})\cr an adtte dataset.}

\item{mae}{(\code{MultiAssayExperiment})\cr contains \code{SummarizedExperiment} objects.}

\item{gene_var}{(\code{character})\cr a gene(s) of interest.}

\item{experiment_name}{(\code{string})\cr the desired experiment to use.}

\item{assay_name}{(\code{string})\cr the assay to define the groups.}
}
\value{
A data frame containing all columns/rows from adtte and select columns from
MAE (assay, Sample IDs) for a given gene(s).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A function to help with merging of MAE to ADTTE object for use with \code{g_km}.
}
\note{
We require that each patient only has one sample.
}
\examples{
mae <- hermes::multi_assay_experiment
adtte <- random.cdisc.data::radtte(cached = TRUE) \%>\% dplyr::mutate(CNSR = as.logical(CNSR))
#make sure patient IDs match some in adtte to test function
experiment_name <- "se2"
se_test <- mae[[experiment_name]]
hd_test <- hermes::HermesData(se_test)
mae_samplemap <- MultiAssayExperiment::sampleMap(mae)
samplemap_experiment <- mae_samplemap[mae_samplemap$assay == experiment_name, ]
se_patients <- samplemap_experiment$primary
adtte$USUBJID[1:9] <- se_patients
gene_var <- c("GeneID:1820", "GeneID:94115")
new_adtte <- h_km_mae_to_adtte(adtte, mae, gene_var = "GeneID:1820", experiment_name = "se2")
new_adtte2 <- h_km_mae_to_adtte(adtte, mae, gene_var = gene_var, experiment_name = "se2")
}
