% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adtteSpec.R
\name{h_km_mae_to_adtte}
\alias{h_km_mae_to_adtte}
\title{Data Preprocessing for \code{ADTTE} Module}
\usage{
h_km_mae_to_adtte(
  adtte,
  mae,
  genes,
  experiment_name = "hd1",
  assay_name = "counts",
  usubjid_var = "USUBJID"
)
}
\arguments{
\item{adtte}{(\verb{data frame})\cr an \code{adtte} dataset.}

\item{mae}{(\code{MultiAssayExperiment})\cr contains \code{AnyHermesData} objects.}

\item{genes}{(\code{GeneSpec})\cr specification for gene(s) (signature), e.g.
using \code{\link[hermes:gene_spec]{hermes::gene_spec()}}.}

\item{experiment_name}{(\code{string})\cr the desired \code{HermesData} to use.}

\item{assay_name}{(\code{string})\cr the assay to define the groups.}

\item{usubjid_var}{(\code{string})\cr variable name of the subject ID variable.}
}
\value{
A data frame containing all columns/rows from \code{adtte} that match
by subject ID with the row names of the MAE and have the gene samples available
in the given experiment. The attribute \code{gene_cols} contains the column names
for the gene columns.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A function to help with merging of MAE to \code{ADTTE}.
}
\note{
The final gene column names can start with a different string than
the original gene IDs (or labels), in particular white space and colons are removed.
}
\examples{
mae <- hermes::multi_assay_experiment
adtte <- teal.modules.hermes::rADTTE \%>\%
  dplyr::mutate(CNSR = as.logical(CNSR))

new_adtte <- h_km_mae_to_adtte(
  adtte,
  mae,
  genes = hermes::gene_spec("GeneID:1820"),
  experiment_name = "hd2"
)
new_adtte2 <- h_km_mae_to_adtte(
  adtte,
  mae,
  genes = hermes::gene_spec(c("GeneID:1820", "GeneID:94115"), fun = colMeans),
  experiment_name = "hd2"
)
new_adtte3 <- h_km_mae_to_adtte(
  adtte,
  mae,
  genes = hermes::gene_spec(c(A = "GeneID:1820", B = "GeneID:94115")),
  experiment_name = "hd2"
)
}
