% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{h_km_mae_to_adtte}
\alias{h_km_mae_to_adtte}
\title{Data Preprocessing for KM Module}
\usage{
h_km_mae_to_adtte(adtte, mae, gene_var, experiment_name, assay_name = "counts")
}
\arguments{
\item{adtte}{(\verb{data frame})\cr an \code{adtte} dataset.}

\item{mae}{(\code{MultiAssayExperiment})\cr contains \code{SummarizedExperiment} objects.}

\item{gene_var}{(\code{character})\cr a gene(s) of interest.}

\item{experiment_name}{(\code{string})\cr the desired experiment to use.}

\item{assay_name}{(\code{string})\cr the assay to define the groups.}
}
\value{
A data frame containing all columns/rows from \code{adtte} and select columns from
MAE (assay, Sample IDs) for a given gene(s).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A function to help with merging of MAE to \code{ADTTE} for use with \code{g_km}.
}
\note{
We require that each patient only has one sample.
}
\examples{
library(dplyr)
library(random.cdisc.data)
mae <- hermes::multi_assay_experiment
adtte <- radtte(cached = TRUE) \%>\%
  mutate(CNSR = as.logical(CNSR))

new_adtte <- h_km_mae_to_adtte(adtte, mae, gene_var = "GeneID:1820", experiment_name = "hd2")
new_adtte2 <- h_km_mae_to_adtte(adtte, mae, gene_var = gene_var, experiment_name = "hd2")
}
