% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSpec.R
\name{h_update_gene_selection}
\alias{h_update_gene_selection}
\title{Helper Function to Update Gene Selection}
\usage{
h_update_gene_selection(session, inputId, selected, choices)
}
\arguments{
\item{session}{(\code{ShinySession})\cr the session object.}

\item{inputId}{(\code{string})\cr
the ID used to call the module input.}

\item{selected}{The new selected value (or multiple values if \code{multiple = TRUE}).
To reset selected value, in case of multiple picker, use \code{character(0)}.}

\item{choices}{List of values to select from. If elements of the list are named
then that name rather than the value is displayed to the user.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This helper function takes the intersection of \code{selected} and
\code{choices} for genes and updates the \code{inputId} accordingly. It then
shows a notification if not all \code{selected} genes were available.
}
