% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_blank}
\alias{is_blank}
\title{Checking for Empty String}
\usage{
is_blank(x)
}
\arguments{
\item{x}{object to check.}
}
\value{
Flag whether \code{x} is identical to an empty string, i.e. \code{""}.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

This predicate function is helpful for functions where arguments could
not yet be initialized from the teal module.
}
\examples{
is_blank("")
is_blank(" ")
}
