% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_convention.R
\name{module_arguments}
\alias{module_arguments}
\title{Standard Module Arguments}
\arguments{
\item{data}{(\code{tdata})\cr
\code{tdata} object which is automatically passed to the UI and server
functions, holding all the data sets provided in the app initialization.}

\item{label}{(\code{string})\cr
menu item label of the module in the teal app.}

\item{inputId}{(\code{string})\cr
the ID used to call the module input.}

\item{adtte_name}{(\code{string})\cr
name of the \code{ADTTE} dataset.}

\item{adtte_vars}{(named \code{list} of \code{string})\cr
names of the variables to use in the \code{ADTTE} dataset. It should comprise elements:
\itemize{
\item \code{aval}: the numeric time-to-event variable.
\item \code{avalu}: the variable holding the unit of \code{aval}.
\item \code{is_event}: the logical event variable. It needs to be \code{TRUE}
when there was an observed event, and \code{FALSE} if the time is censored without
observed event.
\item \code{paramcd}: the character or factor parameter code variable, defining the
type of time-to-event for selection in the module.
\item \code{usubjid}: the subject ID variable.
}}

\item{mae_name}{(\code{string})\cr
name of the MAE data used in the teal module.}

\item{exclude_assays}{(\code{character})\cr
names of the assays which should not be included in choices in the teal module.}

\item{summary_funs}{(named \code{list} of functions or \code{NULL})\cr functions which can be used
in the the gene signatures. For modules that support also multiple genes without
summary, \code{NULL} can be included to not summarize the genes but provide all of them.}

\item{reporter}{(\code{Reporter}) object}

\item{pre_output}{(\code{shiny.tag} or \code{NULL})\cr
placed before the output to put the output into context (for example a title).}

\item{post_output}{(\code{shiny.tag} or \code{NULL})\cr
placed after the output to put the output into context (for example the \code{\link[shiny:helpText]{shiny::helpText()}}
elements can be useful).}

\item{id}{(\code{string}) the shiny module id.}

\item{plot_height}{(\code{list})\cr list of integers to set the default, minimum,
and maximum plot height.}

\item{plot_width}{(\code{list})\cr list of integers to set the default, minimum,
and maximum plot width.}

\item{filter_panel_api}{(\code{FilterPanelAPI})\cr object describing the actual filter panel API.}
}
\description{
The documentation to this function lists all the conventional arguments in
\code{hermes} teal modules.
}
\keyword{internal}
