% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleVarSpec.R
\name{multiSampleVarSpecServer}
\alias{multiSampleVarSpecServer}
\title{Module Server for Specification of Multiple Sample Variables}
\usage{
multiSampleVarSpecServer(inputIds, original_data, ...)
}
\arguments{
\item{inputIds}{(\code{character})\cr multiple input IDs corresponding to the
different sample variables specified in the UI function.}

\item{original_data}{(reactive \code{SummarizedExperiment})\cr input experiment where the
sample variables extracted via \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::colData()}} should be eligible for
selection.}

\item{...}{additional arguments as documented in \code{\link[=sampleVarSpecServer]{sampleVarSpecServer()}},
namely the mandatory \code{experiment_name} and the optional \code{categorical_only},
\code{num_levels} and \code{label_modal_title}.
\code{transformed_data} and \code{assign_lists} should not be
specified as they are already specified internally here.}
}
\value{
List with the final transformed \code{experiment_data} reactive and a
list \code{vars} which contains the selected sample variables as reactives
under their input ID.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

When multiple sample variables are used in a given module, then this
wrapper makes it much easier to specify in the server function.
}
\examples{
\dontrun{
# In the server use:
sample_var_specs <- multiSampleVarSpecServer(
  inputIds = c("facet_var", "color_var"),
  experiment_name = reactive({
    input$experiment_name
  }),
  original_data = <original_data> # nolint Please update the <original_data>
)
# Then can extract the transformed data and selected variables later:
experiment_data <- sample_var_specs$experiment_data()
facet_var <- sample_var_specs$vars$facet_var()
color_var <- sample_var_specs$vars$color_var()
}
}
