% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{teal.modules.hermes-package}
\alias{teal.modules.hermes}
\alias{teal.modules.hermes-package}
\title{\code{teal.modules.hermes} Package}
\description{
\code{teal.modules.hermes} provides RNA-seq analysis modules to add to a
teal application.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/teal.modules.hermes/}
  \item Report bugs at \url{https://github.com/insightsengineering/teal.modules.hermes/issues}
}

}
\author{
\strong{Maintainer}: NEST \email{basel.nestcicd@roche.com}

}
