% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.R
\name{tm_g_forest_tte}
\alias{tm_g_forest_tte}
\alias{ui_g_forest_tte}
\alias{srv_g_forest_tte}
\alias{sample_tm_g_forest_tte}
\title{Teal Module for Survival Forest Plot}
\usage{
tm_g_forest_tte(
  label,
  adtte_name,
  mae_name,
  exclude_assays = "counts",
  summary_funs = list(Mean = colMeans, Median = matrixStats::colMedians, Max =
    matrixStats::colMaxs),
  pre_output = NULL,
  post_output = NULL,
  plot_height = c(600L, 200L, 2000L),
  plot_width = c(1360L, 500L, 2000L)
)

ui_g_forest_tte(id, datasets, mae_name, summary_funs, pre_output, post_output)

srv_g_forest_tte(
  input,
  output,
  session,
  datasets,
  adtte_name,
  mae_name,
  exclude_assays,
  summary_funs,
  plot_height,
  plot_width
)

sample_tm_g_forest_tte()
}
\arguments{
\item{label}{(\code{string})\cr
menu item label of the module in the teal app.}

\item{adtte_name}{(\code{string})\cr
name of the ADTTE dataset.}

\item{mae_name}{(\code{string})\cr
name of the MAE data used in the teal module.}

\item{exclude_assays}{(\code{character})\cr
names of the assays which should not be included in choices in the teal module.}

\item{summary_funs}{(named \code{list} of functions or \code{NULL})\cr functions which can be used
in the the gene signatures. For modules that support also multiple genes without
summary, \code{NULL} can be included to not summarize the genes but provide all of them.}

\item{pre_output}{(\code{shiny.tag} or \code{NULL})\cr
placed before the output to put the output into context (for example a title).}

\item{post_output}{(\code{shiny.tag} or \code{NULL})\cr
placed after the output to put the output into context (for example the \code{\link[shiny:helpText]{shiny::helpText()}}
elements can be useful).}

\item{plot_height}{(\code{list})\cr list of integers to set the default, minimum,
and maximum plot height.}

\item{plot_width}{(\code{list})\cr list of integers to set the default, minimum,
and maximum plot width.}

\item{id}{(\code{string})\cr
the ID used to call the teal module's UI function.}

\item{datasets}{(\code{Datasets})\cr
teal specific argument which is automatically passed to the UI and server
functions, holding all the data sets provided in the app initialization.}

\item{input}{(\code{ReactiveValues})\cr the session's input object.}

\item{output}{(\code{shinyoutput})\cr the session's output object.}

\item{session}{(\code{ShinySession})\cr the session object.}
}
\value{
Shiny module to be used in the teal app.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This module provides an interactive survival forest plot.
}
\section{Functions}{
\itemize{
\item \code{ui_g_forest_tte}: sets up the user interface.

\item \code{srv_g_forest_tte}: sets up the server with reactive graph.

\item \code{sample_tm_g_forest_tte}: sample module function.
}}

\examples{
mae <- hermes::multi_assay_experiment
adtte <- scda::synthetic_cdisc_data("rcd_2021_07_07")$adtte \%>\%
  dplyr::mutate(CNSR = as.logical(.data$CNSR))

data <- teal_data(
  dataset(
    "ADTTE",
    adtte,
    code = 'adtte <- scda::synthetic_cdisc_data("rcd_2021_07_07")$adtte \%>\%
      dplyr::mutate(CNSR = as.logical(.data$CNSR))'
  ),
  dataset("MAE", mae)
)
app <- init(
  data = data,
  modules = root_modules(
    tm_g_forest_tte(
      label = "forestplot",
      adtte_name = "ADTTE",
      mae_name = "MAE"
    )
  )
)
\dontrun{
shinyApp(app$ui, app$server)
}
\dontrun{
# Alternatively you can run the sample module with this function call:
sample_tm_g_forest_tte()
}
}
