% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{tm_g_scatterplot}
\alias{tm_g_scatterplot}
\alias{ui_g_scatterplot}
\alias{srv_g_scatterplot}
\alias{sample_tm_g_scatterplot}
\title{Teal Module for RNA-seq Scatterplot}
\usage{
tm_g_scatterplot(
  label,
  mae_name,
  exclude_assays = "counts",
  summary_funs = list(Mean = colMeans, Median = matrixStats::colMedians, Max =
    matrixStats::colMaxs),
  pre_output = NULL,
  post_output = NULL
)

ui_g_scatterplot(id, data, mae_name, summary_funs, pre_output, post_output)

srv_g_scatterplot(
  id,
  data,
  filter_panel_api,
  reporter,
  mae_name,
  exclude_assays,
  summary_funs
)

sample_tm_g_scatterplot()
}
\arguments{
\item{label}{(\code{string})\cr
menu item label of the module in the teal app.}

\item{mae_name}{(\code{string})\cr
name of the MAE data used in the teal module.}

\item{exclude_assays}{(\code{character})\cr
names of the assays which should not be included in choices in the teal module.}

\item{summary_funs}{(named \code{list} of functions or \code{NULL})\cr functions which can be used
in the the gene signatures. For modules that support also multiple genes without
summary, \code{NULL} can be included to not summarize the genes but provide all of them.}

\item{pre_output}{(\code{shiny.tag} or \code{NULL})\cr
placed before the output to put the output into context (for example a title).}

\item{post_output}{(\code{shiny.tag} or \code{NULL})\cr
placed after the output to put the output into context (for example the \code{\link[shiny:helpText]{shiny::helpText()}}
elements can be useful).}

\item{id}{(\code{string}) the shiny module id.}

\item{data}{(\code{tdata})\cr
\code{tdata} object which is automatically passed to the UI and server
functions, holding all the data sets provided in the app initialization.}

\item{filter_panel_api}{(\code{FilterPanelAPI})\cr object describing the actual filter panel API.}

\item{reporter}{(\code{Reporter}) object}
}
\value{
Shiny module to be used in the teal app.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This module provides an interactive scatterplot for RNA-seq gene expression
analysis.
}
\section{Functions}{
\itemize{
\item \code{ui_g_scatterplot()}: sets up the user interface.

\item \code{srv_g_scatterplot()}: sets up the server with reactive graph.

\item \code{sample_tm_g_scatterplot()}: sample module function.

}}
\examples{
mae <- hermes::multi_assay_experiment
mae_data <- dataset("MAE", mae)
data <- teal_data(mae_data)
app <- init(
  data = data,
  modules = modules(
    tm_g_scatterplot(
      label = "scatterplot",
      mae_name = "MAE"
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

# Alternatively you can run the sample module with this function call:
if (interactive()) {
  sample_tm_g_scatterplot()
}
}
