% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSpec.R
\name{validate_gene_spec}
\alias{validate_gene_spec}
\title{Validation of Gene Specification}
\usage{
validate_gene_spec(gene_spec, gene_choices)
}
\arguments{
\item{gene_spec}{(\code{GeneSpec})\cr gene specification.}

\item{gene_choices}{(\code{character})\cr all possible gene choices.}
}
\description{
This validation function checks that a given \code{\link[hermes:GeneSpec]{hermes::GeneSpec}} has at least
one gene selected and that all genes are included in possible choices.
}
